refactor(services): move helper folder to top level
