+++ src/services/requests/auth/index.ts
@@ -3,7 +3,7 @@ import { fetchAPI } from '@/services/api';
 import { User } from '../users/type';
 import { FetchResult } from '../../type';
 import { createAsyncErrorMessage, withSafeAsync } from '../../utils';
-import { AsyncError } from '../helper/AsyncError';
+import { AsyncError } from '@/services/helpers';
 
 interface LoginData {
   username: string;
