+++ src/services/requests/products/index.ts
@@ -2,7 +2,7 @@ import { fetchAPI } from '@/services/api';
 import { FetchResult } from '../../type';
 import { GetProductsReq, GetProductsRes } from './type';
 import { createAsyncErrorMessage, withSafeAsync } from '../../utils';
-import { AsyncError } from '../helper/AsyncError';
+import { AsyncError } from '@/services/helpers';
 
 export const PRODUCT_TYPES = {
   INGREDIENT: 'ingredientName',
