+++ src/services/utils/withSafeAsync.ts
@@ -1,4 +1,4 @@
-import { AsyncError } from '../requests/helper/AsyncError';
+import { AsyncError } from '@/services/helpers';
 import { FetchResult } from '../type';
 
 function withSafeAsync<T, U extends V[], V>(
